Update getting-started.md
Other than the small edits, I felt this document needs a better structure. With 9 top level items, it is too flat.  At a high level, it has several components: (1) build the kernel with Rust support; (2) install necessary tools in the kernel for writing kernel module in Rust; (3) how to compile and prepare a kernel module in Rust.